+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Expose `RestError` at the crate root.
   * Add an `ImpersonationClient` type to use impersonation tokens with an
     existing client.
+  * Support `/projects/:id/packages` endpoint.
 
 # v0.1508.0